addressed kormat's feedback
